refactor: 경로 단축, 변수명 변경 (#160)
