+++ src/fsd_entities/notification/ui/AllYearToggle.tsx
@@ -1,11 +1,11 @@
 'use client';
 
-interface Props {
+interface AllYearToggleProps {
   checked: boolean;
   onChange: (val: boolean) => void;
 }
 
-export const AllYearToggle = ({ checked, onChange }: Props) => {
+export const AllYearToggle = ({ checked, onChange }: AllYearToggleProps) => {
   return (
     <label className="flex items-center gap-2">
       <span className="text-lg">모든 학번의 경조사 알림 받기</span>
