+++ src/fsd_widgets/notification/index.ts
@@ -1,0 +1,2 @@
+// ui
+export { NotificationSettingWidget } from './ui/NotificationSettingWidget';
